Decomposition9 added

+++ src/Algorithmization/Decomposition9.java
@@ -4,29 +4,29 @@ import java.util.Scanner;
 
 public class Decomposition9 {
     public static void main(String[] args) {
-        int x, y, z, t;
+        double x, y, z, t;
 
         System.out.print("Enter the size of side x: ");
-        x = getInt();
+        x = getDouble();
 
         System.out.print("Enter the size of side y: ");
-        y = getInt();
+        y = getDouble();
 
         System.out.print("Enter the size of side z: ");
-        z = getInt();
+        z = getDouble();
 
         System.out.print("Enter the size of side t: ");
-        t = getInt();
+        t = getDouble();
 
         System.out.println("Square = " + square(x, y, z, t));
     }
 
-    private static double square(int a, int b, int c, int d) {
-        int p = (a + b + c + d) / 2;
+    private static double square(double a, double b, double c, double d) {
+        double p = ((a + b + c + d) / 2);
         return Math.sqrt((p - a) * (p - b) * (p - c) * (p - d));
     }
 
-    private static int getInt() {
-        return new Scanner(System.in).nextInt();
+    private static double getDouble() {
+        return new Scanner(System.in).nextDouble();
     }
 }
